feat: build windows binaries
Currently zot project doesn't build and ship Windows binaries.
This PR adds that support.
