+++ pkg/api/controller.go
@@ -8,10 +8,8 @@ import (
 	"net"
 	"net/http"
 	"os"
-	"runtime"
 	"strconv"
 	"strings"
-	"syscall"
 	"time"
 
 	"github.com/gorilla/mux"
@@ -118,27 +116,6 @@ func NewController(appConfig *config.Config) *Controller {
 	}
 
 	return &controller
-}
-
-func DumpRuntimeParams(log log.Logger) {
-	var rLimit syscall.Rlimit
-
-	evt := log.Info().Int("cpus", runtime.NumCPU()) //nolint: zerologlint
-
-	err := syscall.Getrlimit(syscall.RLIMIT_NOFILE, &rLimit)
-	if err == nil {
-		evt = evt.Uint64("max. open files", uint64(rLimit.Cur)) //nolint: unconvert // required for *BSD
-	}
-
-	if content, err := os.ReadFile("/proc/sys/net/core/somaxconn"); err == nil {
-		evt = evt.Str("listen backlog", strings.TrimSuffix(string(content), "\n"))
-	}
-
-	if content, err := os.ReadFile("/proc/sys/user/max_inotify_watches"); err == nil {
-		evt = evt.Str("max. inotify watches", strings.TrimSuffix(string(content), "\n"))
-	}
-
-	evt.Msg("runtime params")
 }
 
 func (c *Controller) GetPort() int {
